"Implement password encryption and remove test data seeding
Removed data seeding script from 'db/scripts/01_create_tables.sql' as it's not a best practice to have test data hardcoded into the database scripts. Test data should be managed separately.

Introduced BCrypt password encryption to User model to improve security. It ensures that passwords are not stored as plaintext in the database. To support this, spring-security-crypto has also been added to the dependencies in pom.xml."